Rename BodyMarkdown to PascalCase format
